# swagger.yaml
openapi: 3.0.0
info:
  title: Tollgate SaaS API
  description: >
    Production-grade multi-tenant API backend covering the full lifecycle of a SaaS product —
    from org signup through API key issuance, rate limiting, usage metering, and Stripe billing.
  version: 1.0.0
servers:
  - url: https://tollgate-api.onrender.com
    description: Live Production Environment
  - url: http://localhost:3000
    description: Local Development Environment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your 15-minute access token.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Enter your generated API key for rate-limited routes.

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      description: UUID of the organization.
      schema:
        type: string
        format: uuid
      example: "00000000-0000-0000-0000-000000000001"
    Uid:
      name: uid
      in: path
      required: true
      description: UUID of the target user (member).
      schema:
        type: string
        format: uuid
      example: "00000000-0000-0000-0000-000000000002"
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role, self-modification attempted, or demo account is read-only.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotMember:
      description: Caller is not a member of this organization.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Request body failed validation — missing or malformed fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: 'Forbidden'
    ValidationError:
      type: object
      description: >
        Returned when a Zod schema parse fails. The global error handler
        converts ZodError to 400 with a structured `details` array.
      properties:
        error:
          type: string
          example: 'Validation failed'
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: 'refreshToken'
              message:
                type: string
                example: 'Refresh token is required'
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Operation completed successfully'

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    Member:
      type: object
      description: An org member as returned by GET /members.
      properties:
        id:
          type: string
          format: uuid
          description: The user's ID.
          example: '00000000-0000-0000-0000-000000000002'
        email:
          type: string
          format: email
          example: 'user@example.com'
        role:
          type: string
          enum: [owner, admin, member]
          example: 'admin'
        joined_at:
          type: string
          format: date-time
          description: >
            ISO 8601 timestamp of when the user joined the org.
            Also used as the cursor value — pass this field's value as the
            `cursor` query parameter to retrieve the next page of members.
          example: '2025-01-15T10:00:00.000Z'
    MembersPage:
      type: object
      description: >
        Cursor-paginated list of org members. Use `pagination.nextCursor` as the
        `?cursor=` value on the next request to fetch the following page.
        `nextCursor` is `null` when you have reached the last page.
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        pagination:
          type: object
          properties:
            limit:
              type: integer
              description: The page size used for this response.
              example: 20
            nextCursor:
              type: string
              format: date-time
              nullable: true
              description: >
                The `joined_at` timestamp of the last member in this page.
                Pass it as `?cursor=<value>` to fetch the next page.
                `null` means this is the final page — no more members exist.
              example: '2025-01-20T08:30:00.000Z'
    UpdatedMember:
      type: object
      description: >
        The updated org_members row returned after a successful PATCH.
        These are the exact 4 columns from the org_members table (RETURNING *).
      properties:
        user_id:
          type: string
          format: uuid
          example: '00000000-0000-0000-0000-000000000002'
        org_id:
          type: string
          format: uuid
          example: '00000000-0000-0000-0000-000000000001'
        role:
          type: string
          enum: [admin, member]
          example: 'admin'
        joined_at:
          type: string
          format: date-time
          example: '2025-01-15T10:00:00.000Z'

    # ── API Keys ─────────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []
...
paths:
...
  /orgs/{orgId}/members:
    get:
      operationId: listMembers
      tags:
        - Organizations & Members
      summary: List Organization Members
      description: >
        Returns a cursor-paginated list of org members ordered by join date ascending.
        Requires at least `member` role.


        **Cursor pagination** — unlike OFFSET, this implementation is O(1) at any page depth.
        The DB evaluates `WHERE joined_at > $cursor` using the `(org_id, joined_at)` index
        and stops after `limit` rows, regardless of how many members precede the cursor.


        **How to paginate:**
        1. Make a request without `?cursor` to get the first page.
        2. If `pagination.nextCursor` is non-null, pass it as `?cursor=<value>` to get the next page.
        3. Repeat until `pagination.nextCursor` is `null`.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: cursor
          in: query
          required: false
          description: >
            Opaque cursor from the previous page's `pagination.nextCursor`.
            Must be the ISO 8601 `joined_at` timestamp of the last member on the previous page.
            Omit on the first request.
          schema:
            type: string
            format: date-time
          example: '2025-01-15T10:00:00.000Z'
        - name: limit
          in: query
          required: false
          description: Maximum number of members to return per page (1–100, default 20).
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
      responses:
        '200':
          description: Cursor-paginated list of org members ordered by join date ascending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembersPage'
              example:
                data:
                  - id: '00000000-0000-0000-0000-000000000002'
                    email: 'alice@example.com'
                    role: 'owner'
                    joined_at: '2025-01-15T10:00:00.000Z'
                  - id: '00000000-0000-0000-0000-000000000003'
                    email: 'bob@example.com'
                    role: 'member'
                    joined_at: '2025-01-20T08:30:00.000Z'
                pagination:
                  limit: 20
                  nextCursor: '2025-01-20T08:30:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotMember'
  /orgs/{orgId}/members/invite:
    post:
      operationId: inviteMember
      tags:
        - Organizations & Members
      summary: Invite a User to Organization
      description: >
        Adds an existing Tollgate user to the org with the specified role.
        Requires `admin` role or above.

        **Note:** The response only returns a confirmation message — it does not include
        the new member's user ID. Call `GET /orgs/{orgId}/members` after this request
        to retrieve the member's ID for subsequent PATCH or DELETE operations.
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                  example: newuser@example.com
                role:
                  type: string
                  enum: [admin, member]
      responses:
        '201':
          description: User invited and added to org.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: 'newuser@example.com added as member'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No registered user found with that email address.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The user is already a member of this org.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/members/{uid}:
    patch:
      operationId: updateMemberRole
      tags:
        - Organizations & Members
      summary: Update Member Role
      description: >
        Updates a member's role within the org. Requires `owner` role.

        **Two additional guards apply at the service layer:**
        - Owners cannot modify their own role (self-modification is blocked).
        - Owners cannot demote another member who already holds the `owner` role —
          ownership transfer is a separate operation, not achievable via this endpoint.

        Valid target roles are `admin` and `member`. The `owner` role cannot be assigned here.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/Uid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
      responses:
        '200':
          description: Role updated — returns the updated org_members row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedMember'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            Insufficient role (requires owner), self-modification attempted,
            target member holds the owner role (cannot be demoted here), or demo account is read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Target member not found in this org.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeMember
      tags:
        - Organizations & Members
      summary: Remove Member from Organization
      description: >
        Removes a member from the org. Requires `admin` role or above.
        The org owner cannot be removed — ownership must be transferred first.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/Uid'
      responses:
        '200':
          description: Member removed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: 'Member removed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient role (requires admin or above), target is the org owner, or demo account is read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Target member not found in this org.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...